Switch power to use O3 instead of Ofast
Ofast enables possibly unsafe optimizations in addition to O3. This
appears to have been added and then just continually copied into later
Power architectures, and it wasn't included in the CMake build system
when that was introduced.

Replace this with O3 so that the same level of optimization is done by
the compiler.